Migrate to the new nodejs installation script
BREAKING! This commit upgrades nodejs from 14 to 18 as the minimum supported version of the new migration script is 18.

Hopefully this suppresses the 60 sec waiting time that we encounter when we use the deprecated script.
```
$ docker-compose build
...

    => [base 3/7] RUN curl -sL https://deb.nodesource.com/setup_14.x | bash -
                               SCRIPT DEPRECATION WARNING
    ================================================================================
TO AVOID THIS WAIT MIGRATE THE SCRIPT Continuing in 60 seconds (press Ctrl-C to abort) ...
```

An additional package `lsb-release` must be manually installed because it used to be installed as part of deprecated installation script but anymore.